Updated seed for price
